requirements.txt
pytest==7.2.1
selenium==4.8.0
Faker==16.8.1
requests~=2.28.2
python-dateutil~=2.8.2
pandas~=2.0.3